Made the tutorial level better, added more stuff to the gameplay, I will have to make a "help" button on H where it shows the controls and toggle it off by pressing H again

+++ Assets/Scripts/HealthIndicator.cs
@@ -16,6 +16,6 @@ public class HealthIndicator : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        heartAnim.SetInteger("Health", playerVal.currentHealth);
+        heartAnim.SetFloat("Health", playerVal.currentHealth);
     }
 }
